% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb.R
\name{genotype}
\alias{genotype}
\title{genotype range}
\usage{
genotype(avg, sd, direction, foldchange = genotype.foldchange())
}
\arguments{
\item{avg}{Average value of the samples normal data.}

\item{sd}{Standard error of the samples normal data.}
}
\description{
genotype range
}
