% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kb.R
\name{measure.range}
\alias{measure.range}
\title{Measure sample range}
\usage{
measure.range(samples, foldchange = genotype.foldchange())
}
\arguments{
\item{samples}{A numeric vector}
}
\value{
A genotype data model
}
\description{
Measure sample range
}
